Order finish

+++ src/main/java/org/example/springecom/model/dto/OrderItemResponse.java
@@ -1,8 +1,10 @@
 package org.example.springecom.model.dto;
 
+import java.math.BigDecimal;
+
 public record OrderItemResponse(
     String productName,
     int quantity,
-    double totalPrice
+    BigDecimal totalPrice
 ) {
 }
